Rename DataVisPopupPlot to DatastreamPopupPlot
